Update init.el configurations
Adjusted the default frame font size and changed the theme. The command to open current files in WebStorm has been simplified. Updated mu4e settings to include an additional email account and added a new mu4e context for it. Other minor updates include an adjustment in mode line format and a simplification in dired-dwim-target.